replace OhmsTranscriptDisplay with OralHistory::TranscriptComponent
