+++ app/components/work_oh_audio_show_component.html.erb
@@ -167,7 +167,7 @@
 
         <% if has_ohms_transcript? %>
           <div class="tab-pane long-text-line-height" id="ohTranscript" role="tabpanel" aria-labelledby="ohTranscriptTab">
-            <%= OhmsTranscriptDisplay.new(work.oral_history_content.ohms_xml).display %>
+            <%= render OralHistory::TranscriptComponent.new(work.oral_history_content.ohms_xml) %>
           </div>
         <% end %>
